DBZ-7184 Extracted split event handling from streaming source
